Added `user_group` and `user_password` models
Removed the group related methods from the generic `user_model` and updated core code to reflect this change.

`user_password` is empty at the moment, but will be used as the touch point for a more secure passwording system.

Also, renamed the `name` and `display_name` columns of user_group to `slug` and `label` respectively