Changed blog & Shop skin detection logic
Official skins are packaged up as composer packages and should be required by the app as needed.

The skin models will search various directories to find skins and optionally filter out sub directories using a regex.

As a result the getting started skins are no longer required in the core of Nails.